+++ modules/incanter-charts/src/incanter/charts.clj
@@ -1223,7 +1223,6 @@
     chart))
 
 (defmacro candle-stick-plot
-  [& options]
   "Produces a candle stick chart
 
    Options:
@@ -1236,17 +1235,17 @@
     :open Key for accessing open value data (defaults to :open)
     :close Key for accessing close value data (defaults to :close)
     :volume Key for accessing volume data (defaults to :volume). Volume data is optional
-    :title (default 'Candle Chart Plot') main title
+    :title (default 'Candle Stick Plot') main title
     :time-label (default empty)
     :value-label (default empty)
     :legend (default false) prints legend
     :series-label (default empty)
 
    Example:
-   (candle-plot :data <dataset>) ;; uses default mappings so the dataset must have :date, :high, :low, :open, :close and :volume keys
+   (candle-stick-plot :data <dataset>) ;; uses default mappings so the dataset must have :date, :high, :low, :open, :close and :volume keys
 
    ;; more customization
-   (candle-plot
+   (candle-stick-plot
       :data dataset
       :high :HighPrice
       :low :LowPrice
@@ -1257,10 +1256,12 @@
       :time-label \"CoB date\"
       :value-label \"Price\"
       :series-label \"Price time series\"
-      :title \"Price information\")"
+      :title \"Price information\")
 
+"
+  [& options]
   `(let [opts# ~(when options (apply assoc {} options))
-         main-title# (or (:title opts#) "Candle Chart Plot")
+         main-title# (or (:title opts#) "Candle Stick Plot")
          args#
          (concat
           (mapcat #(vector % (or (opts# %) %)) [:volume :high :low :open :close :date])
@@ -1271,7 +1272,6 @@
                               :value-label (or (opts# :value-label) "")
                               :series-label (or (opts# :series-label))]))))]
      (apply candle-stick-plot* args#)))
-
 
 (defn time-series-plot* [x y & options]
   (apply create-xy-series-plot x y create-time-series-plot options))
@@ -1286,7 +1286,7 @@
     :data (default nil) If the :data option is provided a dataset,
                         column names can be used instead of sequences
                         of data as arguments to xy-plot.
-    :title (default 'Time Series Plot') main title
+    :title (default '') main title
     :x-label (default x expression)
     :y-label (default y expression)
     :legend (default false) prints legend
@@ -1317,6 +1317,7 @@
     `(let [opts# ~(when options (apply assoc {} options))
            group-by# (:group-by opts#)
            title# (or (:title opts#) "")
+           legend# (or (:legend opts#) false)
            x-lab# (or (:x-label opts#) (str '~x))
            y-lab# (or (:y-label opts#) (str '~y))
            series-lab# (or (:series-label opts#) (if group-by#
@@ -1325,6 +1326,7 @@
            args# (concat [~x ~y] (apply concat (seq (apply assoc opts#
                                                            [:group-by group-by#
                                                             :title title#
+                                                            :legend legend#
                                                             :x-label x-lab#
                                                             :y-label y-lab#
                                                             :series-label series-lab#]))))]
@@ -1392,7 +1394,7 @@
   to write it to a file.
 
   Options:
-    :title (default 'Histogram') main title
+    :title (default '') main title
     :x-label (default x expression)
     :y-label (default 'Frequency')
     :legend (default false) prints legend
@@ -1860,7 +1862,7 @@
     values -- a sequence of numeric values
 
   Options:
-    :title (default 'Histogram') main title
+    :title (default '') main title
     :x-label (default 'Categories')
     :y-label (default 'Value')
     :legend (default false) prints legend
@@ -2000,7 +2002,7 @@
     values -- a sequence of numeric values
 
   Options:
-    :title (default 'Histogram') main title
+    :title (default '') main title
     :x-label (default 'Categories')
     :y-label (default 'Value')
     :series-label
@@ -2517,7 +2519,7 @@
     values -- a sequence of numeric values
 
   Options:
-    :title (default 'Histogram') main title
+    :title (default '') main title
     :legend (default false) prints legend
 
 
@@ -2607,7 +2609,7 @@
   to write it to a file.
 
   Options:
-    :title (default 'Histogram') main title
+    :title (default '') main title
     :x-label (default x expression)
     :y-label (default 'Frequency')
     :legend (default false) prints legend
@@ -2685,7 +2687,7 @@
   display the chart, or the 'save' function to write it to a file.
 
   Options:
-    :title (default 'Histogram') main title
+    :title (default '') main title
     :x-label (default x expression)
     :y-label (default 'Frequency')
     :legend (default false) prints legend
@@ -3641,6 +3643,21 @@
      (.setSeriesPaint renderer series color)
      chart)))
 
+(defn set-point-size
+  "Set the point size of a scatter plot. Use series option to apply
+  point-size to only one series."
+  [chart point-size & {:keys [series dataset] :or {series :all dataset 0}}]
+  (let [xy (- (/ point-size 2))
+        new-point (java.awt.geom.Ellipse2D$Double. xy xy point-size point-size)
+        plot (.getPlot chart)
+        series-count (.getSeriesCount plot)
+        series-list (if (= :all series)
+                      (range 0 series-count)
+                      (list series))
+        renderer (.getRenderer plot dataset)]
+    (doseq [a-series series-list]
+      (doto renderer (.setSeriesShape a-series new-point)))
+    chart))
 
 ;;;; DEFAULT THEME METHODS